//HOLTERS METHOD UPDATE////// //Circuit.cpp/// -10- Changed the name of T to sampleperiod
-18-
constructor now calls setParams directly

-26-
setup circuit now populates the components, initialise incidents, populates the constant system S0 and then refreshes teh full circuit.

-39-
renamed populateCircuitMatrices to populateComponentMatrices.
This method now populates all the constant circuit elements then calls the refreshPotentiometer method to additionally populate the variable elements

-69-74-
added incidentPot initialisation

-100-
refreshFullCircuit now refreshes the potentiometers, then the statespace, then the nonlin functions. This means that only the variable elements are being refreshed

-107-118-
Added method to refresh the variable circuit elements

-122-
refreshSystemMatrix now populateConstantSystemMatrix.  This is called in the setupMethod and provides S0 for the constant elements of the system matrix

-139-190-
refreshNonLinStateSpace changed to populateConstantStateSpaceTerms().
This method now populates the constant state space terms and is called once when the circuit is created.

-194-213-
refreshStateSpace now refreshes the state space matrices using the constant terms and the new variable terms

-239-286-
Changed how the setters work. They no longer default a value but instead check if the params within range, if it's too high sets it to max, if its too low sets it to min

-298-
set params method is now the public way of setting parameters and calls setFuzz and setVol on the arguements

-316-
Sets the sample rate and samplePeriod then calls the setupCircuit method to refresh the whole circuit

-328-356-
Added some getters to allow for testing.